+++ src/transform/disease.rs
@@ -723,6 +723,8 @@ pub fn from_mydisease_hit(hit: MyDiseaseHit) -> Disease {
         survival_note: None,
         funding: None,
         funding_note: None,
+        diagnostics: None,
+        diagnostics_note: None,
         civic: None,
         disgenet: None,
         xrefs,
@@ -1027,6 +1029,8 @@ mod tests {
             disgenet: None,
             funding: None,
             funding_note: None,
+            diagnostics: None,
+            diagnostics_note: None,
             xrefs: HashMap::new(),
         };